# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Kaordo API
  version: 0.0.1
servers:
  - url: http://localhost:8081
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak access token with aud=kerno-api
    internalToken:
      type: apiKey
      in: header
      name: X-Kaordo-Internal-Token
      description: Derived token shared only by the local Kerno and Nodo processes.
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          description: A stable, non-sensitive reason when an authentication request is rejected.
...
    NodoUpload:
      type: object
      required: [id, kind, mimeType, filename, width, height, size, complete]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [image, video, file] }
        mimeType: { type: string }
        filename: { type: string }
        width: { type: integer }
        height: { type: integer }
        size: { type: integer, format: int64 }
        complete: { type: boolean }
...
paths:
  /v1/uploads/{id}/meta:
    servers:
      - url: http://localhost:8082
    get:
      operationId: getNodoUploadMetadata
      description: The owning account can read processing state for a resumable upload.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Processed upload metadata, ready for an authorized application reference.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NodoUpload' }
        '202':
          description: Upload or media processing is not complete.
        default:
          description: Invalid, expired, failed or inaccessible upload.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
...